# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ORVELL PULSE Wholesale ERP Backend API",
        "description": "Production REST API documentation for ORVELL PULSE Wholesale ERP system with Spatie RBAC, multi-company isolation, inventory tracking, cashier settlement, and automated reporting.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Default API Host"
        }
    ],
    "paths": {
...
        "/api/invoices/amendment-request": {
            "post": {
                "tags": [
                    "Invoice Amendment"
                ],
                "summary": "Submit request to amend an immutable finalized invoice",
                "description": "Submit Invoice Amendment Request",
                "operationId": "630ce753990bd4be2f8da5e5ba22de64",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "invoice_id",
                                    "reason",
                                    "requested_changes"
                                ],
                                "properties": {
                                    "invoice_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "reason": {
                                        "type": "string",
                                        "example": "Price discount approved by manager"
                                    },
                                    "requested_changes": {
                                        "type": "object",
                                        "example": {
                                            "discount_amount": 25
                                        }
                                    },
                                    "company_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Invoice amendment request submitted successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/invoices/amendment/{id}/approve": {
            "post": {
                "tags": [
                    "Invoice Amendment"
                ],
                "summary": "Approve invoice amendment request and generate versioned invoice",
                "description": "Approve Invoice Amendment Request (Admin / Manager Only)",
                "operationId": "72a537691d225ff29da54b10fd1a4055",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Amendment Request ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "company_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Invoice amendment approved and new version created"
                    },
                    "403": {
                        "description": "Unauthorized. Only Admins and Managers can approve"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/invoices/amendment/{id}/reject": {
            "post": {
                "tags": [
                    "Invoice Amendment"
                ],
                "summary": "Reject invoice amendment request with stated reason",
                "description": "Reject Invoice Amendment Request (Admin / Manager Only)",
                "operationId": "5836195e4adbc9ef5615a2ffd1264a52",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Amendment Request ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "rejection_reason"
                                ],
                                "properties": {
                                    "rejection_reason": {
                                        "type": "string",
                                        "example": "Amendment rejected due to missing physical bale verification"
                                    },
                                    "company_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Invoice amendment request rejected successfully"
                    },
                    "403": {
                        "description": "Unauthorized. Only Admins and Managers can reject"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/invoices/amendment/list": {
            "get": {
                "tags": [
                    "Invoice Amendment"
                ],
                "summary": "List invoice amendment requests for company",
                "description": "List Invoice Amendment Requests",
                "operationId": "59270394c681977c3931f8c2b16e2b10",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by request status (pending, approved, rejected)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "pending"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Amendment requests retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter Sanctum personal access token as: Bearer <token>",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}